Manual

API Reference

Quickhull.quickhull — Function

quickhull(points, options=Quickhull.Options())

Compute the convex hull of points. points can be a vector of point-like objects (e.g. Tuple or StaticVector) or a (D, N)-sized matrix of numbers.

See documentation for Quickhull.Options.

Quickhull.delaunay — Function

delaunay(points, options=Quickhull.Options())

Compute the d-dimensional Delaunay triangulation of points. points can be a vector of point-like objects (e.g. Tuple or StaticVector) or a (D, N)-sized matrix of numbers.

The triangulation is found by lifting into (d+1) dimensions and taking the convex hull.

Quickhull.Options — Type

Quickhull.Options(options...)

Avaliable options are:

kernel – a subtype of HyperplaneKernel used for hyperplane calculations. HyperplaneKernelExactSIMD by default.

indextype – a subtype of Integer that specifies how vertex indices should be stored. Int32 by default.

joggle – whether to joggle the input points. false by default

joggle_amount – how much to joggle the input points. 100.0 by default.

statistics – whether to record statistics. false by default.

subdivide – controls whether hull is computed by subdividing the input points and merging the resulting sub-hulls. Available options are:

NoSubdivide() – don't use subdivision (default).

SerialSubdivide(chunks=nchunks, levels=nlevels) – subdivide points into nchunks many chunks, nlevels many times recursively. Not parallel.

ParallelSubdivide(chunks=nchunks, levels=nlevels) – subdivide points into nchunks many chunks, nlevels many times recursively. Sub-hulls are computed in parallel.

Quickhull.points — Method

points(hull)

The points the hull was constructed from. This includes points inside the hull - see vertices(hull) and vertexpoints(hull).

Quickhull.vertices — Method

vertices(hull)

The indices of points that are vertices of the hull.

Quickhull.facets — Method

facets(hull)

The facets of the hull. A facet is defined by D vertices.

GeometryBasics.Mesh — Method

GeometryBasics.Mesh(hull::Quickhull.AbstractHull)

Create a Mesh from the points and facets of hull.

Quickhull.PolyhedraLibrary — Function

PolyhedraLibrary(solver)

Create an instance of a Polyhedra.Library with the given solver that uses quickhull as a backend. Requires the Polyhedra package to be loaded.